requirements.txt
flask
pymupdf
python-docx
reportlab
